[#671] renaming almost finished: check contents
